input과 list 파일 분리 생성
Input, InputContainer, List, ListContainer로 관심사를 분리하여 파일을 생성하였습니다
