+++ src/List.jsx
@@ -1,0 +1,7 @@
+export default function List() {
+  return (
+    <li>
+      리스트
+    </li>
+  );
+}
