+++ src/ListContainer.jsx
@@ -1,0 +1,7 @@
+import List from './List';
+
+export default function ListContainer() {
+  return (
+    <List />
+  );
+}
